renamed MauiSample to MauiSampleApp, cause i constantly confused which project was the lib and which the app
