+++ AvaloniaMauiHybrid.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.3.32922.545
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiSample", "MauiSample\MauiSample.csproj", "{D2634557-8A3C-4062-9480-3C6A5A4CC147}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avalonia.Maui", "Avalonia.Maui\Avalonia.Maui.csproj", "{CD6CAB48-9C52-48CC-90BC-46D6F079F0CF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MauiSampleApp", "MauiSampleApp\MauiSampleApp.csproj", "{6F55A510-234E-48CF-92B3-338E89E5A954}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,16 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Debug|Any CPU.Deploy.0 = Debug|Any CPU
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D2634557-8A3C-4062-9480-3C6A5A4CC147}.Release|Any CPU.Deploy.0 = Release|Any CPU
 		{CD6CAB48-9C52-48CC-90BC-46D6F079F0CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{CD6CAB48-9C52-48CC-90BC-46D6F079F0CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD6CAB48-9C52-48CC-90BC-46D6F079F0CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CD6CAB48-9C52-48CC-90BC-46D6F079F0CF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F55A510-234E-48CF-92B3-338E89E5A954}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F55A510-234E-48CF-92B3-338E89E5A954}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F55A510-234E-48CF-92B3-338E89E5A954}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F55A510-234E-48CF-92B3-338E89E5A954}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
