+++ MauiSampleApp/App.xaml
@@ -1,8 +1,8 @@
 ﻿<?xml version = "1.0" encoding = "UTF-8" ?>
 <Application xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:local="clr-namespace:MauiSample"
-             x:Class="MauiSample.App">
+             xmlns:local="clr-namespace:MauiSampleApp"
+             x:Class="MauiSampleApp.App">
     <Application.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
