+++ MauiSampleApp/App.xaml.cs
@@ -1,4 +1,4 @@
-﻿namespace MauiSample
+﻿namespace MauiSampleApp
 {
     public partial class App : Application
     {
