+++ MauiSampleApp/AppShell.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <Shell
-    x:Class="MauiSample.AppShell"
+    x:Class="MauiSampleApp.AppShell"
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:local="clr-namespace:MauiSample"
-    xmlns:pages="clr-namespace:MauiSample.Avalonia.Views"
+    xmlns:local="clr-namespace:MauiSampleApp"
+    xmlns:pages="clr-namespace:MauiSampleApp.Avalonia.Views"
     Shell.FlyoutBehavior="Flyout">
 
     <ShellContent
