+++ MauiSampleApp/AvaloniaApp.axaml.cs
