+++ MauiSampleApp/Platforms/Android/AndroidManifest.xml
