+++ MauiSampleApp/Platforms/Android/MainActivity.cs
