+++ MauiSampleApp/Platforms/Android/MainApplication.cs
