+++ MauiSampleApp/Platforms/Android/Resources/values/colors.xml
