+++ MauiSampleApp/Platforms/iOS/AppDelegate.cs
