+++ MauiSampleApp/Platforms/iOS/Program.cs
