+++ MauiSampleApp/Resources/Fonts/OpenSans-Regular.ttf
