+++ MauiSampleApp/Resources/Fonts/OpenSans-Semibold.ttf
