+++ MauiSampleApp/Resources/Raw/AboutAssets.txt
