+++ MauiSampleApp/Resources/Styles/Colors.xaml
